Add files via upload
This project is a simple flappy bird game made in C++. Just need to fly between the obstacles like a column of pipes without hitting them.
In this I have used a 2-D character array for making the bird. The player can control the 2-D bird by using the space key.